nail-salon-ai-saas/backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
aiomysql==0.2.0
pymssql==2.2.11
alembic==1.12.1
psycopg2-binary==2.9.9

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Redis
redis==5.0.1
hiredis==2.2.3

# Background tasks
celery==5.3.4
flower==2.0.1

# Vanna and AI
vanna==0.3.4
ollama==0.1.6
chromadb==0.4.18

# ML and Analytics
scikit-learn==1.3.2
prophet==1.1.5
pandas==2.1.3
numpy==1.26.2

# Monitoring and logging
prometheus-client==0.19.0
sentry-sdk==1.38.0

# Utilities
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2
python-dateutil==2.8.2
chardet==5.2.0
aiofiles==23.2.1

